Fix to PWMRANGE
The analogWrite-Range is no longer 1023 by default. It has been changed to 255.
"NOTE: The default analogWrite range was 1023 in releases before 3.0, but this lead to incompatibility with external libraries which depended on the Arduino core default of 256. Existing applications which rely on the prior 1023 value may add a call to analogWriteRange(1023) to their setup() routine to return to their old behavior." 
https://arduino-esp8266.readthedocs.io/en/latest/reference.html?highlight=analogwrite#analog-output